+++ .gitignore
@@ -7,3 +7,5 @@ node_modules
 package-lock.json
 yarn.lock
 .vscode
+
+interop/*-image.json
