--- interop/.gitignore
@@ -1,5 +1,0 @@
-dist
-node-image.json
-chromium-image.json
-firefox-image.json
-webkit-image.json
