+++ interop/BrowserDockerfile
@@ -1,15 +1,25 @@
 # syntax=docker/dockerfile:1
 
+# Copied since we won't have the repo to use if expanding from cache.
+
+# Workaround: https://github.com/docker/cli/issues/996
+ARG BASE_IMAGE=node-js-libp2p-head
+FROM ${BASE_IMAGE} as js-libp2p-base
+
 FROM mcr.microsoft.com/playwright
 
-COPY --from=node-js-libp2p-head /app/ /app/
-WORKDIR /app/interop
+COPY --from=js-libp2p-base /app/ /app/
+WORKDIR /app
+# We install browsers here instead of the cached version so that we use the latest browsers at run time.
+# Ideally this would also be pinned, but playwright controls this, so there isn't much we can do about it.
+# By installing here, we avoid installing it at test time.
+RUN npm ci
+RUN npx playwright install-deps
+RUN npx playwright install
+RUN npm run build
+
 # Options: chromium, firefox, webkit
 ARG BROWSER=chromium
-ENV BROWSER=$BROWSER
-# disable colored output and CLI animation from test runners
-ENV CI true
-ENV DEBUG libp2p*
+ENV BROWSER=${BROWSER}
 
-# manually specify runner until https://github.com/hugomrdias/playwright-test/issues/572 is resolved
-ENTRYPOINT npm run test:interop:multidim -- --build false --types false -t browser -- --browser $BROWSER
+ENTRYPOINT npm test -- -t browser -- --browser $BROWSER
