--- interop/chromium-version.json
@@ -1,21 +1,0 @@
-{
-  "id": "chromium-js-libp2p-head",
-  "containerImageID": "chromium-js-libp2p-head",
-  "transports": [
-    {
-      "name": "webtransport",
-      "onlyDial": true
-    },
-    {
-      "name": "webrtc-direct",
-      "onlyDial": true
-    },
-    "webrtc",
-    {
-      "name": "wss",
-      "onlyDial": true
-    }
-  ],
-  "secureChannels": ["noise"],
-  "muxers": ["mplex", "yamux"]
-}
