--- interop/firefox-version.json
@@ -1,17 +1,0 @@
-{
-  "id": "firefox-js-libp2p-head",
-  "containerImageID": "firefox-js-libp2p-head",
-  "transports": [
-    {
-      "name": "webrtc-direct",
-      "onlyDial": true
-    },
-    "webrtc",
-    {
-      "name": "wss",
-      "onlyDial": true
-    }
-  ],
-  "secureChannels": ["noise"],
-  "muxers": ["mplex", "yamux"]
-}
