--- interop/test/ping.spec.ts
@@ -1,254 +1,0 @@
-/* eslint-disable no-console */
-/* eslint-env mocha */
-
-import { } from 'aegir/chai'
-import { noise } from '@chainsafe/libp2p-noise'
-import { yamux } from '@chainsafe/libp2p-yamux'
-import { circuitRelayTransport } from '@libp2p/circuit-relay-v2'
-import { type Identify, identify } from '@libp2p/identify'
-import { mplex } from '@libp2p/mplex'
-import { ping, type PingService } from '@libp2p/ping'
-import { tcp } from '@libp2p/tcp'
-import { webRTC, webRTCDirect } from '@libp2p/webrtc'
-import { webSockets } from '@libp2p/websockets'
-import * as filters from '@libp2p/websockets/filters'
-import { webTransport } from '@libp2p/webtransport'
-import { type Multiaddr, multiaddr } from '@multiformats/multiaddr'
-import { createLibp2p, type Libp2p, type Libp2pOptions } from 'libp2p'
-
-async function redisProxy (commands: any[]): Promise<any> {
-  const res = await fetch(`http://localhost:${process.env.REDIS_PROXY_PORT}`, {
-    method: 'POST',
-    body: JSON.stringify(commands)
-  })
-
-  if (!res.ok) {
-    throw new Error('Redis command failed')
-  }
-
-  return res.json()
-}
-
-let node: Libp2p<{ ping: PingService, identify: Identify }>
-const isDialer: boolean = process.env.is_dialer === 'true'
-const timeoutSecs: string = process.env.test_timeout_secs ?? '180'
-
-describe('ping test', function () {
-  // make the default timeout longer than the listener timeout
-  this.timeout((parseInt(timeoutSecs) * 1000) + 30000)
-
-  // eslint-disable-next-line complexity
-  beforeEach(async () => {
-    // Setup libp2p node
-    const TRANSPORT = process.env.transport
-    const SECURE_CHANNEL = process.env.security
-    const MUXER = process.env.muxer
-    const IP = process.env.ip ?? '0.0.0.0'
-
-    const options: Libp2pOptions<{ ping: PingService, identify: Identify }> = {
-      start: true,
-      connectionManager: {
-        minConnections: 0
-      },
-      connectionGater: {
-        denyDialMultiaddr: async () => false
-      },
-      services: {
-        ping: ping(),
-        identify: identify()
-      }
-    }
-
-    switch (TRANSPORT) {
-      case 'tcp':
-        options.transports = [tcp()]
-        options.addresses = {
-          listen: isDialer ? [] : [`/ip4/${IP}/tcp/0`]
-        }
-        break
-      case 'webtransport':
-        options.transports = [webTransport()]
-        if (!isDialer) {
-          throw new Error('WebTransport is not supported as a listener')
-        }
-        break
-      case 'webrtc-direct':
-        options.transports = [webRTCDirect()]
-        options.addresses = {
-          listen: isDialer ? [] : [`/ip4/${IP}/udp/0/webrtc-direct`]
-        }
-        break
-      case 'webrtc':
-        options.transports = [webRTC(),
-          webSockets({ filter: filters.all }), // ws needed to connect to relay
-          circuitRelayTransport({
-            discoverRelays: 1
-          }) // needed to use the relay
-        ]
-        options.addresses = {
-          listen: isDialer ? [] : ['/webrtc']
-        }
-        options.connectionGater = {
-          denyDialMultiaddr: () => {
-            // by default we refuse to dial local addresses from the browser since they
-            // are usually sent by remote peers broadcasting undialable multiaddrs but
-            // here we are explicitly connecting to a local node so do not deny dialing
-            // any discovered address
-            return false
-          }
-        }
-        break
-      case 'ws':
-        options.transports = [webSockets()]
-        options.addresses = {
-          listen: isDialer ? [] : [`/ip4/${IP}/tcp/0/ws`]
-        }
-        break
-      case 'wss':
-        process.env.NODE_TLS_REJECT_UNAUTHORIZED = '0'
-        options.transports = [webSockets()]
-        options.addresses = {
-          listen: isDialer ? [] : [`/ip4/${IP}/tcp/0/wss`]
-        }
-        break
-      default:
-        throw new Error(`Unknown transport: ${TRANSPORT ?? '???'}`)
-    }
-
-    let skipSecureChannel = false
-    let skipMuxer = false
-    switch (TRANSPORT) {
-      case 'webtransport':
-      case 'webrtc-direct':
-        skipSecureChannel = true
-        skipMuxer = true
-        break
-      case 'webrtc':
-        skipSecureChannel = true
-        skipMuxer = true
-        // Setup yamux and noise to connect to the relay node
-        options.streamMuxers = [yamux()]
-        options.connectionEncryption = [noise()]
-        break
-      default:
-        // Do nothing
-    }
-
-    if (!skipSecureChannel) {
-      switch (SECURE_CHANNEL) {
-        case 'noise':
-          options.connectionEncryption = [noise()]
-          break
-        default:
-          throw new Error(`Unknown secure channel: ${SECURE_CHANNEL ?? ''}`)
-      }
-    }
-
-    if (!skipMuxer) {
-      switch (MUXER) {
-        case 'mplex':
-          options.streamMuxers = [mplex()]
-          break
-        case 'yamux':
-          options.streamMuxers = [yamux()]
-          break
-        default:
-          throw new Error(`Unknown muxer: ${MUXER ?? '???'}`)
-      }
-    }
-
-    node = await createLibp2p(options)
-  })
-
-  afterEach(async () => {
-    // Shutdown libp2p node
-    try {
-      // We don't care if this fails
-      await node.stop()
-    } catch { }
-  });
-
-  // eslint-disable-next-line complexity
-  (isDialer ? it.skip : it)('should listen for ping', async () => {
-    try {
-      const sortByNonLocalIp = (a: Multiaddr, b: Multiaddr): number => {
-        if (a.toString().includes('127.0.0.1')) {
-          return 1
-        }
-        return -1
-      }
-      let multiaddrs = node.getMultiaddrs().sort(sortByNonLocalIp).map(ma => ma.toString())
-
-      const transport = process.env.transport
-      if (transport === 'webrtc') {
-        const relayAddr = process.env.RELAY_ADDR
-        const hasWebrtcMultiaddr = new Promise<string[]>((resolve) => {
-          const abortController = new AbortController()
-          node.addEventListener('self:peer:update', (event) => {
-            const webrtcMas = node.getMultiaddrs().filter(ma => ma.toString().includes('/webrtc'))
-            if (webrtcMas.length > 0) {
-              resolve(webrtcMas.sort(sortByNonLocalIp).map(ma => ma.toString()))
-            }
-            abortController.abort()
-          }, { signal: abortController.signal })
-        })
-
-        if (relayAddr == null || relayAddr === '') {
-          throw new Error('No relayAddr')
-        }
-        // const conn = await node.dial(multiaddr(relayAddr))
-        console.error('dial relay')
-        await node.dial(multiaddr(relayAddr))
-        console.error('wait for relay reservation')
-        multiaddrs = await hasWebrtcMultiaddr
-      }
-
-      console.error('inform redis of dial address')
-      // Send the listener addr over the proxy server so this works on both the Browser and Node
-      await redisProxy(['RPUSH', 'listenerAddr', multiaddrs[0]])
-      // Wait
-      console.error('wait for incoming ping')
-      await new Promise(resolve => setTimeout(resolve, 1000 * parseInt(timeoutSecs, 10)))
-    } catch (err) {
-      // Show all errors in an aggregated error
-      if (err instanceof AggregateError) {
-        console.error('unexpected exception in ping test Errors:', err.errors)
-      } else {
-        console.error('unexpected exception in ping test:', err)
-      }
-      throw err
-    }
-  });
-
-  // eslint-disable-next-line complexity
-  (isDialer ? it : it.skip)('should dial and ping', async () => {
-    try {
-      let otherMaStr: string = (await redisProxy(['BLPOP', 'listenerAddr', timeoutSecs]).catch(err => { throw new Error(`Failed to wait for listener: ${err}`) }))[1]
-      // Hack until these are merged:
-      // - https://github.com/multiformats/js-multiaddr-to-uri/pull/120
-      otherMaStr = otherMaStr.replace('/tls/ws', '/wss')
-
-      const otherMa = multiaddr(otherMaStr)
-      const handshakeStartInstant = Date.now()
-
-      console.error(`node ${node.peerId.toString()} dials: ${otherMa}`)
-      await node.dial(otherMa)
-
-      console.error(`node ${node.peerId.toString()} pings: ${otherMa}`)
-      const pingRTT = await node.services.ping.ping(multiaddr(otherMa))
-      const handshakePlusOneRTT = Date.now() - handshakeStartInstant
-      console.log(JSON.stringify({
-        handshakePlusOneRTTMillis: handshakePlusOneRTT,
-        pingRTTMilllis: pingRTT
-      }))
-    } catch (err) {
-      // Show all errors in an aggregated error
-      if (err instanceof AggregateError) {
-        console.error('unexpected exception in ping test Errors:', err.errors)
-      } else {
-        console.error('unexpected exception in ping test:', err)
-      }
-      throw err
-    }
-  })
-})
