--- interop/webkit-version.json
@@ -1,12 +1,0 @@
-{
-    "id": "webkit-js-libp2p-head",
-    "containerImageID": "webkit-js-libp2p-head",
-    "transports": [],
-    "secureChannels": [
-        "noise"
-    ],
-    "muxers": [
-        "mplex",
-        "yamux"
-    ]
-}
